requirements.txt
pillow
streamlit
requests
boto3
